Finished maybe add/edit, view Item, expanded search, tags dialog (#54)
* added triangle class that extends shape

* added colour to shape class

* added some attributes to add_edit before my computer broke

added description, serial number and half date. It broke when I tried adding a date type instead of regular text :(

* Reformatted and added an icon

tried to keep going where I left off but android studio broke again when I tried to add another icon

* almost done add/edit just needs tags

If it wasn't for android studio breaking on me I would make way less frequent commits

* It broke again

I lost a lot of progress on that last commit so I had to redo some stuff here but it doesn't look like all the changes I made this time will be committed either :(

* add/edit UI done + my android studio updated and changed things in build.gradle.kts

hopefully if everyone updates we won't have a problem ?

* Last little edits to add/edit item

added mock code for tags to be view as a ChipGroup

* image visibility + add tag button

I kept thinking of stuff as I was writing the description of the pull request tehe

* finished Item View

tried to make it as close to the mockup as possible

* Added expanded search to activity_main.xml

Added sort by and filter by options
Scrap having to pull down to see expanded for this sprint, it will always show

* added tag dialog layout

I keep using Chip Group for tags and I sure hope it works. If it isn't the easiest implementation of tags we can change it and blame chat GPT :)

---------

Co-authored-by: Matthew Jacobs <89949279+WindowsDotExe@users.noreply.github.com>