Big tags stuff  (#68)
* added triangle class that extends shape

* added colour to shape class

* added some attributes to add_edit before my computer broke

added description, serial number and half date. It broke when I tried adding a date type instead of regular text :(

* Reformatted and added an icon

tried to keep going where I left off but android studio broke again when I tried to add another icon

* almost done add/edit just needs tags

If it wasn't for android studio breaking on me I would make way less frequent commits

* It broke again

I lost a lot of progress on that last commit so I had to redo some stuff here but it doesn't look like all the changes I made this time will be committed either :(

* add/edit UI done + my android studio updated and changed things in build.gradle.kts

hopefully if everyone updates we won't have a problem ?

* Last little edits to add/edit item

added mock code for tags to be view as a ChipGroup

* image visibility + add tag button

I kept thinking of stuff as I was writing the description of the pull request tehe

* finished Item View

tried to make it as close to the mockup as possible

* Added expanded search to activity_main.xml

Added sort by and filter by options
Scrap having to pull down to see expanded for this sprint, it will always show

* added tag dialog layout

I keep using Chip Group for tags and I sure hope it works. If it isn't the easiest implementation of tags we can change it and blame chat GPT :)

* I started working on tags but I wanna move to deleting items

Literally one of the main points of our meeting  :dizzy_face:

* more tags regret

* Failed to implement deleting but it launches with no error and it's basically pseudo code

I'm really tired

* Last attempt of using checkboxes for selecting

I'll try doing just multiple choice slection

* I tried really hard to delete items :(

* Deleting items is really hard

* Added log messages to try and find the problem

* Added delete functionality

instead of flipping between add and deleting fabs I just put them next to each other. That's next sprint problem

* Add some comments and docs to all files

* All tests ran for ItemList

* Added delete test and docs

* tags probably

* Delete TagDialog.java

* fixed wrong merge

* added tagdialog and click listner

* Merge branch 'main' into LinaSoup

* old commit to get rid of

* Tags dialog is on the way

it opens and clicks and takes input, there is just a layout bug that isn't showing the tag but android studio is freezing so imma commit then restart to fix it

* added more ui and toast messages

* Item now has existing tags and selected tags

still need to get tags with a db so imma attempt that now

* back to working on multiple selection cuz db stuff is hard

only single selection is working for some reason

* Multiple selection works

now to work on being able to pass around two TagLists, one for selected items and one for the global existing tags that the user can pick from and add/delete  from

* created tag manger to handle existing and selected tags

I over complicated TagList and tried to make it do a lot of things but it wasn't really going to work

* Decided on db setup for tags

more details in pull request
also added update selected tags for the item class and add tags for tagslist class. both are db versions of tag implementation

* Add db version of delete tag from item

As I am making these, I'm commenting where to put them in the Tag dialog once db is merged

* Delete TagManager.java

* removed old double tagList implementaion

---------

Co-authored-by: Matthew Jacobs <89949279+WindowsDotExe@users.noreply.github.com>
Co-authored-by: RylannAnd <100882433+RylannAnd@users.noreply.github.com>